requirements.txt
streamlit>=1.45.1
pandas>=1.3.0
numpy>=1.21.0
plotly>=5.3.1
google-cloud-storage>=2.11.0
